examples/MNIST_RAW_format/requirements.txt
tensorflow-macos
kafka-python